Improve image conversion and splitting
- Refactor the .czi to .tiff conversion function using opencv to save
them in grayscale images
- Add function for loading a .tiff file
- Rename directory for saving the chunks of an image
- Add imagecodecs requirement, necessary for the czifile module
- Typos and renaming
