+++ requirements.txt
@@ -1,5 +1,16 @@
+cycler==0.10.0
 czifile==2019.7.2
+imagecodecs==2020.5.30
+kiwisolver==1.3.1
+matplotlib==3.3.3
+mpmath==1.1.0
 numpy==1.19.2
+opencv-python==4.4.0.46
+Pillow==8.0.1
+pkg-resources==0.0.0
+pyparsing==2.4.7
+python-dateutil==2.8.1
+pytz==2020.1
 six==1.15.0
 sympy==1.6.2
 tifffile==2020.10.1
